feat(views): fill the home page with the provider feed
