+++ crates/music/src/youtube/client.rs
@@ -8,8 +8,8 @@ use ytmusic::YtMusic;
 
 use crate::youtube::{genres, wire};
 use crate::{
-    Album, AlbumDetail, Artist, ArtistProfile, Genre, GenreDetail, MediaKind, MusicApi, Playlist,
-    PlaylistDetail, SavedArtist, Track, UserProfile,
+    Album, AlbumDetail, Artist, ArtistProfile, Genre, GenreDetail, GenreSection, MediaKind,
+    MusicApi, Playlist, PlaylistDetail, SavedArtist, Track, UserProfile,
 };
 
 const PORTRAIT_LIMIT: usize = 24;
@@ -315,6 +315,10 @@ impl MusicApi for YouTubeClient {
             .enumerate()
             .map(|(index, track)| wire::track(track, index as u32))
             .collect())
+    }
+
+    async fn home(&self) -> Result<Vec<GenreSection>> {
+        genres::home(&self.api).await
     }
 
     async fn genres(&self) -> Result<Vec<Genre>> {
