+++ crates/views/src/shared/mod.rs
@@ -9,6 +9,7 @@ pub(crate) mod menu;
 pub(crate) mod page;
 pub(crate) mod picks;
 pub(crate) mod playlist_editor;
+pub(crate) mod shelves;
 pub(crate) mod tracks;
 
 use gpui::prelude::*;
